refactor: split EmojiDataCollector.kt into two collectors
